feat: Add user picture to session data
This commit adds the user's picture to the session data returned by the GetSession endpoint in the auth controller. The picture field is now included in the userData map and is returned as part of the sessionData JSON response.

fix: Handle error when creating user

This commit fixes an issue in the CreateUser endpoint in the users controller. Previously, when there was an error creating a user in the database, a generic error message was returned. Now, the specific error message returned by the database is included in the response.

feat: Update user table schema

This commit updates the user table schema in the database. It adds a new column called "picture" of type text to store the URL of the user's profile picture.